four main operator commit

+++ include/Accumulator/Accumulator.h
@@ -12,7 +12,7 @@ public:
 
     Accumulator(const Accumulator &other) = default;
 
-    int TaskDefinition(const std::vector<int> &Storage, int StartIndex, int EndIndex) override;
+    double TaskDefinition(const std::vector<double> &Storage, int StartIndex, int EndIndex) override;
 
 };
