+++ include/Subtractor/Subtractor.h
@@ -10,7 +10,7 @@ class Subtractor : public Operator {
 public:
     Subtractor();
 
-    int TaskDefinition(const std::vector<int> &Storage, int StartIndex, int EndIndex) override;
+    double TaskDefinition(const std::vector<double> &Storage, int StartIndex, int EndIndex) override;
 };
 
 #endif //SUBTRACTOR_H
